fix(room-ui):Ajuste no erro Participant/Checkin que estavam duplicados, agora são únicos em model, CheckInDao + findByRa, layouts/manifest/cores e cadastro/check-in alinhados ao Room

+++ app/src/main/java/com/br/techweekmobile/dao/ProjectDao.java
@@ -26,5 +26,5 @@ public interface ProjectDao {
     LiveData<List<Project>> getAll();
 
     @Query("SELECT * FROM project WHERE participant_id = :participantId")
-    LiveData<List<Project>> getByParticipant(int participantId);
+    LiveData<List<Project>> getByParticipant(long participantId);
 }
